OpenAI_agent_SDK/project1/pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
